+++ lib/src/paths/lib.dart
@@ -1,4 +1,4 @@
 export 'my_paths.dart';
-export 'my_user_data_directory_store.dart';
-export 'my_user_data_directory_validator.dart';
-export 'my_user_data_files_migrator.dart';
+export 'my_user_data_dir_session.dart';
+export 'my_user_data_dir_store.dart';
+export 'my_user_data_dir_validator.dart';